emacs: remove unused features: tags, java stacktrace linking to files
I have never actually used those features, and making them work has
been a hassle: the tags support requires tags.io.  The java stacktrace
index is slow to build and very hacky.

A simpler and cleaner alternative to the previous stacktrace
implementation would be: add a custom command to open a file in the
projectd indicated by a line in the stack trace.  That is very easy to
do by converting the full class name to path and grep in the
SRC_FILES.